bakendo/requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==3.6.4.0
boto3==1.26.142
botocore==1.29.142
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2.1
django-celery==3.1.17
djangorestframework==3.14.0
jmespath==1.0.1
kombu==5.2.4
prompt-toolkit==3.0.38
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
